fix: resolving merge conflict
