+++ Scribe/AppTexts/AppTextStyling.swift
@@ -81,3 +81,38 @@ func getGlobeIcon(fontSize: CGFloat) -> NSAttributedString {
 
   return NSAttributedString(attachment: globeAttachment)
 }
+
+/// Formats and returns the text of the Scribe privacy policy with links activated.
+func setPrivacyPolicy(fontSize: CGFloat, title: String, text: String) -> NSMutableAttributedString {
+  let privacyTextTitle = NSMutableAttributedString(
+    string: title, attributes: [NSAttributedString.Key.font: UIFont.boldSystemFont(ofSize: fontSize * 1.5)]
+  )
+
+  let wikidataDataLicensing: String = "https://www.wikidata.org/wiki/Wikidata:Licensing"
+  let wikipediaDataLicensing: String = "https://en.wikipedia.org/wiki/Wikipedia:Reusing_Wikipedia_content"
+  let huggingFaceLicensing: String = "https://github.com/huggingface/transformers/blob/master/LICENSE"
+  let scribeGitHub: String = "https://github.com/scribe-org"
+  let scribeEmail: String = "scribe.langauge@gmail.com"
+  let gitHubLogoLicensing: String = "https://github.com/logos"
+  let wikidataLogoLicensing: String = "https://foundation.wikimedia.org/wiki/Policy:Trademark_policy"
+  let customKeyboardLicense: String = "https://github.com/EthanSK/CustomKeyboard/blob/master/LICENSE"
+
+  let privacyPolicyTextWithLinks = addHyperLinks(
+    originalText: text,
+    links: [
+      wikidataDataLicensing: wikidataDataLicensing,
+      wikipediaDataLicensing: wikipediaDataLicensing,
+      huggingFaceLicensing: huggingFaceLicensing,
+      scribeGitHub: scribeGitHub,
+      scribeEmail: "mailto:" + scribeEmail,
+      gitHubLogoLicensing: gitHubLogoLicensing,
+      wikidataLogoLicensing: wikidataLogoLicensing,
+      customKeyboardLicense: customKeyboardLicense],
+    fontSize: fontSize
+  )
+
+  return concatAttributedStrings(
+    left: privacyTextTitle,
+    right: privacyPolicyTextWithLinks
+  )
+}
